Removed the usage of --no-sound-null-safety, updated version, changelog
